+++ examples/config-example-3.js
@@ -5,7 +5,7 @@
 "use strict";
 
 var path = require("path"),
-	opt = require("./opt").create();
+	opt = require("opt").create();
 
 var config = { name: "fred", "email": "fred@example.com" },
 	search_paths = [ "config-example-1.conf",