+++ examples/json2sh.js
@@ -12,7 +12,7 @@
 
 var fs = require("fs"),
     path = require("path"),
-    opt = require("./opt").create(),
+    opt = require("opt").create(),
     input = "",
     output = "",
     fields = {},